misc: remove public API
